icecream-qr/requirements.txt
pandas
qrcode
cryptography
python-dotenv
Pillow
pyzbar
opencv-python
streamlit
streamlit-webrtc
aiortc
av

